Add: XRTC1.0-封装Tcp连接类

+++ src/server/signaling_server.h
@@ -54,7 +54,7 @@ private:
 
     int _listen_fd = -1;
     std::vector<SignalingWorker*> _workers;
-    int _next_worker_index = 0;
+    size_t _next_worker_index = 0;
 };
 
 } // namespace xrtc
